backend/requirements.txt
# Django Core
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5

# Database
# psycopg2-binary==2.9.9  # PostgreSQL (for production)

# Cache & Tasks
redis==5.0.1
celery==5.3.4
celery[redis]==5.3.4

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Payments
stripe==7.8.0

# Image Processing
Pillow==10.1.0

# HTTP & APIs
requests==2.31.0
httpx==0.25.2

# AI/ML - LiteLLM Integration
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2

# Utilities
python-slugify==8.0.1
python-dateutil==2.8.2
pytz==2023.3

# Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.39.1

# Development
black==23.12.1
flake8==7.0.0
pytest==7.4.3
pytest-django==4.7.0
